rename files to reflect function name
